+++ idf_build_apps/main.py
@@ -2,14 +2,12 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import argparse
-import io
 import json
 import os
 import re
 import shutil
 import sys
 import textwrap
-import warnings
 from pathlib import (
     Path,
 )
@@ -55,22 +53,22 @@ def _check_app_dependency(
     manifest_rootpath,  # type: str
     modified_components,  # type: list[str] | None
     modified_files,  # type: list[str] | None
-    ignore_component_dependencies_file_patterns,  # type: list[str] | None
+    ignore_app_dependencies_filepatterns,  # type: list[str] | None
 ):  # type: (...) -> bool
     # not check since modified_components and modified_files are not passed
     if modified_components is None and modified_files is None:
         return False
 
-    # not check since ignore_component_dependencies_file_patterns is passed and matched
+    # not check since ignore_app_dependencies_filepatterns is passed and matched
     if (
-        ignore_component_dependencies_file_patterns
+        ignore_app_dependencies_filepatterns
         and modified_files is not None
-        and files_matches_patterns(modified_files, ignore_component_dependencies_file_patterns, manifest_rootpath)
+        and files_matches_patterns(modified_files, ignore_app_dependencies_filepatterns, manifest_rootpath)
     ):
         LOGGER.debug(
             'Skipping check component dependencies for apps since files %s matches patterns: %s',
             ', '.join(modified_files),
-            ', '.join(ignore_component_dependencies_file_patterns),
+            ', '.join(ignore_app_dependencies_filepatterns),
         )
         return False
 
@@ -95,7 +93,7 @@ def find_apps(
     default_build_targets=None,  # type: list[str] | str | None
     modified_components=None,  # type: list[str] | str | None
     modified_files=None,  # type: list[str] | str | None
-    ignore_component_dependencies_file_patterns=None,  # type: list[str] | str | None
+    ignore_app_dependencies_filepatterns=None,  # type: list[str] | str | None
     sdkconfig_defaults=None,  # type: str | None
 ):  # type: (...) -> list[App]
     """
@@ -139,9 +137,9 @@ def find_apps(
     :type modified_components: list[str] | str | None
     :param modified_files: modified files
     :type modified_files: list[str] | str | None
-    :param ignore_component_dependencies_file_patterns: file patterns that used for ignoring checking the component
+    :param ignore_app_dependencies_filepatterns: file patterns that used for ignoring checking the component
         dependencies
-    :type ignore_component_dependencies_file_patterns: list[str] | str | None
+    :type ignore_app_dependencies_filepatterns: list[str] | str | None
     :param sdkconfig_defaults: semicolon-separated string, pass to idf.py -DSDKCONFIG_DEFAULTS if specified,
         also could be set via environment variables "SDKCONFIG_DEFAULTS"
     :type sdkconfig_defaults: str | None
@@ -166,7 +164,7 @@ def find_apps(
 
     modified_components = to_list(modified_components)
     modified_files = to_list(modified_files)
-    ignore_component_dependencies_file_patterns = to_list(ignore_component_dependencies_file_patterns)
+    ignore_app_dependencies_filepatterns = to_list(ignore_app_dependencies_filepatterns)
 
     apps = []
     if target == 'all':
@@ -195,7 +193,7 @@ def find_apps(
                         manifest_rootpath=manifest_rootpath,
                         modified_components=modified_components,
                         modified_files=modified_files,
-                        ignore_component_dependencies_file_patterns=ignore_component_dependencies_file_patterns,
+                        ignore_app_dependencies_filepatterns=ignore_app_dependencies_filepatterns,
                     ),
                     modified_components=modified_components,
                     modified_files=modified_files,
@@ -223,7 +221,7 @@ def build_apps(
     manifest_rootpath=None,  # type: str | None
     modified_components=None,  # type: list[str] | str | None
     modified_files=None,  # type: list[str] | str | None
-    ignore_component_dependencies_file_patterns=None,  # type: list[str] | str | None
+    ignore_app_dependencies_filepatterns=None,  # type: list[str] | str | None
 ):  # type: (...) -> (int, list[App]) | int
     """
     Build all the specified apps
@@ -258,9 +256,9 @@ def build_apps(
     :type modified_components: list[str] | str | None
     :param modified_files: modified files
     :type modified_files: list[str] | str | None
-    :param ignore_component_dependencies_file_patterns: file patterns that used for ignoring checking the component
+    :param ignore_app_dependencies_filepatterns: file patterns that used for ignoring checking the component
         dependencies
-    :type ignore_component_dependencies_file_patterns: list[str] | str | None
+    :type ignore_app_dependencies_filepatterns: list[str] | str | None
     :return: (exit_code, built_apps) if specified ``modified_components``
     :rtype: int, list[App]
     :return: exit_code if not specified ``modified_components``
@@ -269,7 +267,7 @@ def build_apps(
     apps = to_list(apps)  # type: list[App]
     modified_components = to_list(modified_components)
     modified_files = to_list(modified_files)
-    ignore_component_dependencies_file_patterns = to_list(ignore_component_dependencies_file_patterns)
+    ignore_app_dependencies_filepatterns = to_list(ignore_app_dependencies_filepatterns)
 
     ignore_warnings_regexes = []
     if ignore_warning_strs:
@@ -300,27 +298,13 @@ def build_apps(
         app.parallel_count = parallel_count
 
         if collect_app_info:
-            if isinstance(collect_app_info, io.TextIOWrapper):
-                warnings.warn(
-                    '"collect_app_info" does not support file stream in idf-build-apps 1.0.0, Please use str instead',
-                    DeprecationWarning,
-                )
-                app._collect_app_info = collect_app_info.name
-            else:
-                app._collect_app_info = collect_app_info
+            app._collect_app_info = collect_app_info
 
             if app.collect_app_info not in collect_files:
                 collect_files.append(app.collect_app_info)
 
         if collect_size_info:
-            if isinstance(collect_size_info, io.TextIOWrapper):
-                warnings.warn(
-                    '"collect_size_info" does not support file stream in idf-build-apps 1.0.0, Please use str instead',
-                    DeprecationWarning,
-                )
-                app._collect_size_info = collect_size_info.name
-            else:
-                app._collect_size_info = collect_size_info
+            app._collect_size_info = collect_size_info
 
             if app.collect_size_info not in collect_files:
                 collect_files.append(app.collect_size_info)
@@ -352,7 +336,7 @@ def build_apps(
                 modified_components=modified_components,
                 modified_files=modified_files,
                 check_app_dependencies=_check_app_dependency(
-                    manifest_rootpath, modified_components, modified_files, ignore_component_dependencies_file_patterns
+                    manifest_rootpath, modified_components, modified_files, ignore_app_dependencies_filepatterns
                 ),
             )
         except BuildError as e:
@@ -468,15 +452,7 @@ class IdfBuildAppsCliFormatter(argparse.HelpFormatter):
             else:
                 default_type = type(action.default)
 
-            if isinstance(action, argparse._AppendAction):  # noqa
-                _help += (
-                    '. Could be specified for multiple times'
-                    '{} ! DeprecationWarning: will change to space-separated list in idf-build-apps 1.0.0 version'.format(
-                        self.LINE_SEP
-                    )
-                )
-                _type = 'list[{}]'.format(default_type.__name__)
-            elif action.nargs in [argparse.ZERO_OR_MORE, argparse.ONE_OR_MORE]:
+            if action.nargs in [argparse.ZERO_OR_MORE, argparse.ONE_OR_MORE]:
                 _type = 'list[{}]'.format(default_type.__name__)
             else:
                 _type = default_type.__name__
@@ -524,7 +500,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
     )
     common_args.add_argument(
         '--exclude',
-        action='append',
+        nargs='+',
         help='Ignore specified directory (if --recursive is given)',
     )
     common_args.add_argument(
@@ -548,7 +524,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
     )
     common_args.add_argument(
         '--config',
-        action='append',
+        nargs='+',
         help='Adds configurations (sdkconfig file names) to build. '
         'This can either be FILENAME[=NAME] or FILEPATTERN. FILENAME is the name of the sdkconfig file, '
         'relative to the project directory, to be used. Optional NAME can be specified, '
@@ -578,7 +554,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
     )
     common_args.add_argument(
         '--manifest-file',
-        action='append',
+        nargs='+',
         help='Manifest files which specify the build test rules of the apps',
     )
     common_args.add_argument(
@@ -590,10 +566,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
         '--default-build-targets',
         nargs='+',
         help='space-separated list of supported targets. Targets supported in current ESP-IDF branch '
-        '(except preview ones) would be used if this option is not set.'
-        '{} ! DeprecationWarning: comma-separated list support will be removed in idf-build-apps 1.0.0 version'.format(
-            IdfBuildAppsCliFormatter.LINE_SEP
-        ),
+        '(except preview ones) would be used if this option is not set.',
     )
     common_args.add_argument(
         '--modified-components',
@@ -612,7 +585,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
     )
     common_args.add_argument(
         '-if',
-        '--ignore-component-dependencies-file-patterns',
+        '--ignore-app-dependencies-filepatterns',
         nargs='*',
         default=None,
         help='space-separated list which specifies the file patterns used for ignoring the component dependencies. '
@@ -675,7 +648,7 @@ def get_parser():  # type: () -> argparse.ArgumentParser
     )
     build_parser.add_argument(
         '--ignore-warning-str',
-        action='append',
+        nargs='+',
         help='Ignore the warning string that match the specified regex in the build output',
     )
     build_parser.add_argument(
@@ -712,20 +685,17 @@ def validate_args(parser, args):  # type: (argparse.ArgumentParser, argparse.Nam
     default_build_targets = []
     if args.default_build_targets:
         for target in args.default_build_targets:
-            t_list = [_t.strip() for _t in target.split(',')] if ',' in target else [target.strip()]
-            for _t in t_list:
-                if _t not in ALL_TARGETS:
-                    raise InvalidCommand(
-                        'Unrecognizable target {} specified with "--default-build-targets". '
-                        'Current ESP-IDF available targets: {}'.format(_t, ALL_TARGETS)
-                    )
+            if target not in ALL_TARGETS:
+                raise InvalidCommand(
+                    'Unrecognizable target {} specified with "--default-build-targets". '
+                    'Current ESP-IDF available targets: {}'.format(target, ALL_TARGETS)
+                )
 
-                if _t not in default_build_targets:
-                    default_build_targets.append(_t)
-
+            if target not in default_build_targets:
+                default_build_targets.append(target)
     args.default_build_targets = default_build_targets
 
-    if args.ignore_component_dependencies_file_patterns:
+    if args.ignore_app_dependencies_filepatterns:
         if args.modified_files is None:
             raise InvalidCommand(
                 'Must specify "--ignore-component-dependencies-file-patterns" with "--modified-files", '
@@ -767,7 +737,7 @@ def main():
         default_build_targets=args.default_build_targets,
         modified_components=args.modified_components,
         modified_files=args.modified_files,
-        ignore_component_dependencies_file_patterns=args.ignore_component_dependencies_file_patterns,
+        ignore_app_dependencies_filepatterns=args.ignore_app_dependencies_filepatterns,
         sdkconfig_defaults=args.sdkconfig_defaults,
     )
 
@@ -801,7 +771,7 @@ def main():
         manifest_rootpath=args.manifest_rootpath,
         modified_components=args.modified_components,
         modified_files=args.modified_files,
-        ignore_component_dependencies_file_patterns=args.ignore_component_dependencies_file_patterns,
+        ignore_app_dependencies_filepatterns=args.ignore_app_dependencies_filepatterns,
     )
 
     if args.modified_components is not None: