+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ## Added
 
 - Support keyword `depends_filepatterns` in the manifest file
+- Support expanding environment variables in the manifest files
 
 ## BREAKING CHANGES
 
@@ -23,6 +24,13 @@ All notable changes to this project will be documented in this file.
 - CLI Options Renamed
   - `--depends-on-components` renamed to `--modified-components`
   - `--depends-on-files` renamed to `--modified-files`
+  - `--ignore-components-dependencies-file-patterns` renamed to `--ignore-app-dependencies-filepatterns`
+- Removed the deprecated CLI call methods, now these options only support space-separated list
+  - `--exclude`
+  - `--config`
+  - `--manifest-file`
+  - `--ignore-warning-str`
+  - `--default-build-targets`
 
 ## [0.6.1] (2023-05-10)